S15.d.3: extract up.ts AgentMesh deploy phase (1296 → 1182 LOC)
Third sub-slice of the up.ts multi-PR sub-train. Extract the
AgentMesh infrastructure deploy phase (Inspektor Gadget eBPF,
relay + registry deployment in-cluster, external-registry
shortcut, optional AGIC ingress for public endpoints) to
cli/src/commands/up/agentmesh_deploy.ts.

Caller in up.ts shrinks to a 9-line dispatch:
  const { deployAgentMesh } = await import("./up/agentmesh_deploy.js");
  const meshResult = await deployAgentMesh(
    { repoRoot, acr, acrLoginServer, baseName, rg, stepper },
    { globalRegistry: options.globalRegistry, exposeRegistry: options.exposeRegistry },
  );
  const registryMode = meshResult.registryMode;
  const globalRegistryUrl = meshResult.globalRegistryUrl;
  const globalRelayUrl = meshResult.globalRelayUrl;

The result triple flows forward unchanged to the ClawSandbox CR
creation step (env-var injection) and saveContext() at end-of-deploy.

Verification:
- up.ts: 1296 → 1182 LOC (further reduction in d.4)
- new up/agentmesh_deploy.ts: 174 LOC
- tsc --noEmit / lint (27 warnings, baseline) / build: clean
- npm test: 454 passed; 2 skipped pre-existing

The redundant outer-scope `const fs = await import("fs")` in
up.ts is removed (it was only used by the AgentMesh manifest
substitution block, which is now scoped to the new module).

Sub-slice d.4 (sandbox bring-up + summary) will continue toward
the §15 800 cap.

No behavioral change: AgentMesh deploy body moved verbatim.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-up-cli-d3.md
